requirements.txt
interactions
python-dotenv
api
